requirements.txt
pyvis
networkx
flask
pyyaml
pyinstaller
selenium
pillow
